refactor: move url parsing to jina clients helper
