+++ jina/clients/helper.py
@@ -1,11 +1,13 @@
 """Helper functions for clients in Jina."""
 
+from argparse import Namespace
 from functools import wraps
-from typing import Callable, Optional
+from typing import Any, Callable, Dict, Optional, Tuple
 
 from jina.excepts import BadClientCallback
-from jina.helper import get_rich_console
+from jina.helper import ArgNamespace, get_rich_console
 from jina.logging.logger import JinaLogger
+from jina.parsers import set_client_cli_parser
 from jina.proto import jina_pb2
 from jina.types.request.data import Response
 
@@ -105,3 +107,83 @@ def callback_exec_on_error(
         on_error(resp, exception)
 
     _safe_callback(on_error_wrap, False, logger)(response)
+
+
+def parse_client(kwargs) -> Namespace:
+    """
+    Parse the kwargs for the Client
+
+    :param kwargs: kwargs to be parsed
+
+    :return: parsed argument.
+    """
+    kwargs = _parse_kwargs(kwargs)
+    return ArgNamespace.kwargs2namespace(
+        kwargs, set_client_cli_parser(), warn_unknown=True
+    )
+
+
+def _parse_kwargs(kwargs: Dict[str, Any]) -> Dict[str, Any]:
+    if 'host' in kwargs.keys():
+        return_scheme = dict()
+        (
+            kwargs['host'],
+            return_scheme['port'],
+            return_scheme['protocol'],
+            return_scheme['tls'],
+        ) = _parse_host_scheme(kwargs['host'])
+
+        for key, value in return_scheme.items():
+            if value:
+                if key in kwargs:
+                    raise ValueError(
+                        f"You can't have two definitions of {key}: you have one in the host scheme and one in the keyword argument"
+                    )
+                elif value:
+                    kwargs[key] = value
+
+    kwargs = _add_default_port_tls(kwargs)
+    kwargs = _delete_host_slash(kwargs)
+
+    return kwargs
+
+
+def _add_default_port_tls(kwargs: Dict[str, Any]) -> Dict[str, Any]:
+    if ('tls' in kwargs) and ('port' not in kwargs):
+        kwargs['port'] = 443
+    return kwargs
+
+
+def _delete_host_slash(kwargs: Dict[str, Any]) -> Dict[str, Any]:
+    if 'host' in kwargs:
+        if kwargs['host'][-1] == '/':
+            kwargs['host'] = kwargs['host'][:-1]
+    return kwargs
+
+
+def _parse_host_scheme(host: str) -> Tuple[str, str, str, bool]:
+    scheme, _hostname, port = _parse_url(host)
+
+    tls = None
+    if scheme in ('grpcs', 'https', 'wss'):
+        scheme = scheme[:-1]
+        tls = True
+
+    if scheme == 'ws':
+        scheme = 'websocket'
+
+    return _hostname, port, scheme, tls
+
+
+def _parse_url(host):
+    if '://' in host:
+        scheme, host = host.split('://')
+    else:
+        scheme = None
+
+    if ':' in host:
+        host, port = host.split(':')
+    else:
+        port = None
+
+    return scheme, host, port
